created a dedicated module for client management
create two files: snpes_client_manager.c/h that holds client management
functions.

also created a dedicated unit test file to validate this module

fix #9
